+++ plugins/wiggle/src/LinearWiggleDisplay/renderSvg.tsx
@@ -1,11 +1,12 @@
 import React from 'react'
 import { getContainingView } from '@jbrowse/core/util'
 import { when } from 'mobx'
-import YScaleBar from '../../shared/YScaleBar'
 import {
   ExportSvgDisplayOptions,
   LinearGenomeViewModel,
 } from '@jbrowse/plugin-linear-genome-view'
+// locals
+import YScaleBar from '../shared/YScaleBar'
 import { WiggleDisplayModel } from './model'
 
 export async function renderSvg(
