+++ plugins/wiggle/src/MultiLinearWiggleDisplay/components/ColorLegend.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { observer } from 'mobx-react'
 
 // locals
-import { WiggleDisplayModel } from '../models/model'
+import { WiggleDisplayModel } from '../model'
 import RectBg from './RectBg'
 
 const ColorLegend = observer(function ({
