+++ plugins/wiggle/src/MultiLinearWiggleDisplay/components/ScoreLegend.tsx
@@ -4,7 +4,7 @@ import { LinearGenomeViewModel } from '@jbrowse/plugin-linear-genome-view'
 import { observer } from 'mobx-react'
 
 // locals
-import { WiggleDisplayModel } from '../models/model'
+import { WiggleDisplayModel } from '../model'
 import RectBg from './RectBg'
 
 type LGV = LinearGenomeViewModel
