+++ plugins/wiggle/src/MultiLinearWiggleDisplay/components/YScaleBars.tsx
@@ -4,7 +4,7 @@ import { LinearGenomeViewModel } from '@jbrowse/plugin-linear-genome-view'
 import { observer } from 'mobx-react'
 
 // locals
-import { WiggleDisplayModel } from '../models/model'
+import { WiggleDisplayModel } from '../model'
 import YScaleBar from '../../shared/YScaleBar'
 import ColorLegend from './ColorLegend'
 import ScoreLegend from './ScoreLegend'
